Add Sign Finder Feature (#15)
* Add Sign Finder feature
* Refactor SignESP and SignFinder to use WorldSign

+++ src/main/java/me/bobulo/mine/pdam/util/BlockPosition.java
@@ -14,8 +14,8 @@ public class BlockPosition {
         return new BlockPosition(pos.getX(), pos.getY(), pos.getZ());
     }
 
-    double x;
-    double y;
-    double z;
+    int x;
+    int y;
+    int z;
 
 }
